+++ R/data.R
@@ -4,7 +4,7 @@
 #'  in the U.S. from 2004 to 2007.  This is a subset of the dataset used in Callaway and
 #'  Sant'Anna (2021).  See that paper for additional descriptions.
 #'
-#' @format A data frame with 2000 rows and 5 variables:
+#' @format A data frame with 2000 rows and 6 variables:
 #' \describe{
 #'   \item{year}{the year of the observation}
 #'   \item{countyreal}{a unique identifier for a particular county}
@@ -16,5 +16,5 @@
 #'    period.}
 #'   \item{treat}{whether or not a particular county is treated in that year}
 #' }
-#' @source Callaway and Sant'Anna (2020)
+#' @source Callaway and Sant'Anna (2021)
 "mpdta"